Fix Pencil Icon Scaling
Hello, I'm pretty new to all of this so apologies if I'm going about it wrong. I saw that the icon kept clipping when I zoomed in, and multiplied when I zoomed out. Did a bit of googling and added that line to my obsidian snippets and it fixed it.